Bugfix for formatting on Import aliases
Newlines were being added to the suffix on an import statement, which
led to the import actually having two newlines in its (combined) suffix.

Also added a 'space_around' to distinguish from 'statement', because
statement has a more specific python meaning. Aliases are not
statements, but they are also not parenthesizable expressions.