Extend `scope` parsing to allow scoping an attribute
`scope` is useful for handling parens around a node, e.g, in
`a + (b - c)`, the second BinOp's `prefix` and `suffix` will be sure
to have the correct parens (and they will be matched). However in some
cases, as in `ImportFrom`, the list of child nodes in `names` can also
be wrapped in parens. Extending `scope` to work on an attribute of the
node as well, or (by default) on the whole node.