Neatly format code that lacks formatting information
Anything that is added to the ast is unlikely to have formatting info
attached to it. In these cases, provide a sane default format to put
whitespaces and newlines, etc, where they belong.

This is probably not perfect yet, but a big improvement over how it
previously handled these cases.

+ a few formatting fixes where needed