Fix call_arguments to not require position information. (#49)
Fix call_arguments to not require position information.

This patch uses getattr to safely access lineno/col_offset. It also fixed the traversal to never produce code containing syntax errors.

The Call printing sorts args and kwargs together, so if one is not careful with col_offset and the like, the printer may write code like this: f(a=0, b), which is illegal.

The change to the sorting function only affects cases where the col_offset is inconsistent (e.g. after manual edits), so it's safe in terms of reversibility. But the changed sorting is nicer in cases where you want to add an arg to an existing tree in a legal manner (but you may not care where exactly it lives): you can now add a keyword arg without having to set its lineno/col_offset, and it is guaranteed to appear in a legal position.