Bugfix for setting indentation level on child nodes without formatting data
Fixes #62

This previously worked by chance by carrying on the "indent diff" (the
amount that a node is indented relative to its container) onto a block's
children if the children are missing formatting info. However, if the
container is at the top level, it doesn't have any indentation to pass
on. It should be inferring this diff from a child node instead.