Fix continued with indentation
Now with indentation handled explicitly, the previous practice of indenting
once per with, even in the continued case, is shown to be buggy. This might
have shown up in a test that tests automatic indentation after a continued
with, but we have no such test.

This commit fixes the visitation such that a continued with statement is
visited without an indented wrapper.