Fix error in end_{line,col} tracking (#100)
* Fix error in end_{line,col} tracking

The current implementation sets the end line and column of a node to the end values for the next token after visiting that node.  However, since visiting a node consumes all of its tokens, this sets them to the end of the first token outside the node.  This makes it seem like the node contains the token after it.

For example, in the test added in this commit for "a = b + c", the current implementation has a's end_col as 3 and b's as 7, which means a contains the "=" and b contains the "+".

This commit fixes the issue by setting the end values to the start values for the next token.